feat: 정산 로직 구현

+++ backend/src/main/java/site/solsoltrip/backend/entity/IndividualWithdraw.java
@@ -22,5 +22,5 @@ public class IndividualWithdraw {
     @ManyToOne(fetch = FetchType.LAZY)
     private AccompanyMemberWithdraw accompanyMemberWithdraw;
 
-    private Integer individual;
+    private Double individual;
 }
